requirements.txt
uvicorn
fastapi
pydantic
pandas
langchain_community
langchain
python-dotenv
numpy
openai
chromadb
requests
langchain_openai
python-multipart
datetime
pathlib
